requirements.txt
joblib
numpy
pandas
python-dotenv
scikit-learn
fastapi
uvicorn
pydantic